refactor: isUserAuthenticated 사용자 검증 JwtAuthenticationFilter로 이동
