+++ src/main/java/com/github/Hyun_jun_Lee0811/dictionary/security/JwtAuthenticationFilter.java
@@ -1,5 +1,8 @@
 package com.github.Hyun_jun_Lee0811.dictionary.security;
 
+import static com.github.Hyun_jun_Lee0811.dictionary.type.ErrorCode.USER_NOT_AUTHENTICATED;
+
+import com.github.Hyun_jun_Lee0811.dictionary.expection.ErrorResponse;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -28,7 +31,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
   protected void doFilterInternal(
       @NonNull HttpServletRequest request,
       @NonNull HttpServletResponse response,
-      @NonNull  FilterChain filterChain) throws ServletException, IOException {
+      @NonNull FilterChain filterChain) throws ServletException, IOException {
 
     String token = resolveToken(request);
     if (StringUtils.hasText(token) && jwtTokenProvider.validateToken(token)) {
@@ -36,6 +39,15 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
       SecurityContextHolder.getContext().setAuthentication(authentication);
     }
     filterChain.doFilter(request, response);
+  }
+
+  public boolean isUserAuthenticated(String username) {
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    if (authentication == null || !authentication.isAuthenticated()) {
+      throw new ErrorResponse(USER_NOT_AUTHENTICATED);
+    }
+
+    return username.equals(authentication.getName());
   }
 
   private String resolveToken(HttpServletRequest request) {
